[modify] Follow Function

+++ src/main/kotlin/kr/kro/tripsketch/services/FollowService.kt
@@ -13,30 +13,51 @@ class FollowService(
     private val jwtService: JwtService,
     private val userRepository: UserRepository
 ) {
+
     fun follow(token: String, following: String) {
         val follower = jwtService.getEmailFromToken(token)
+        if (follower == following) {
+            throw IllegalArgumentException("자신을 구독할 수 없습니다.")
+        }
         userService.findUserByEmail(following) ?: throw IllegalArgumentException("사용자가 존재하지 않습니다.")
         if (!followRepository.existsByFollowerAndFollowing(follower, following)) {
             followRepository.save(Follow(follower = follower, following = following))
+        } else {
+            throw IllegalArgumentException("이미 구독 중입니다.")
         }
     }
 
     fun unfollow(token: String, following: String) {
         val follower = jwtService.getEmailFromToken(token)
+        if (follower == following) {
+            throw IllegalArgumentException("자신을 구독 취소할 수 없습니다.")
+        }
         userService.findUserByEmail(following) ?: throw IllegalArgumentException("사용자가 존재하지 않습니다.")
-        followRepository.deleteByFollowerAndFollowing(follower, following)
+        if (followRepository.existsByFollowerAndFollowing(follower, following)) {
+            followRepository.deleteByFollowerAndFollowing(follower, following)
+        } else {
+            throw IllegalArgumentException("구독 하지 않은 사용자를 취소 할 수 없습니다.")
+        }
     }
 
     fun unfollowMe(token: String, follower: String) {
         val following = jwtService.getEmailFromToken(token)
+        if (follower == following) {
+            throw IllegalArgumentException("자신을 구독 취소를 할 수 없습니다.")
+        }
         userService.findUserByEmail(follower) ?: throw IllegalArgumentException("사용자가 존재하지 않습니다.")
-        followRepository.deleteByFollowerAndFollowing(follower, following)
+        if (followRepository.existsByFollowerAndFollowing(follower, following)) {
+            followRepository.deleteByFollowerAndFollowing(follower, following)
+        } else {
+            throw IllegalArgumentException("구독 하지 않은 사용자를 취소할 수 없습니다.")
+        }
     }
 
     fun getFollowings(followerEmail: String): List<UserProfileDto> {
         val follower = userRepository.findByEmail(followerEmail)
         val followerProfile = follower?.let {
             UserProfileDto(
+                email = it.email,
                 nickname = it.nickname,
                 introduction = it.introduction,
                 profileImageUrl = it.profileImageUrl
@@ -46,6 +67,7 @@ class FollowService(
         val followings = followRepository.findByFollower(followerEmail).map { follow ->
             val user = userRepository.findByEmail(follow.following)
             UserProfileDto(
+                email = user?.email,
                 nickname = user?.nickname,
                 introduction = user?.introduction,
                 profileImageUrl = user?.profileImageUrl
@@ -60,6 +82,7 @@ class FollowService(
         val following = userRepository.findByEmail(followingEmail)
         val followingProfile = following?.let {
             UserProfileDto(
+                email = it.email,
                 nickname = it.nickname,
                 introduction = it.introduction,
                 profileImageUrl = it.profileImageUrl
@@ -69,6 +92,7 @@ class FollowService(
         val followers = followRepository.findByFollowing(followingEmail).map { follow ->
             val user = userRepository.findByEmail(follow.follower)
             UserProfileDto(
+                email = user?.email,
                 nickname = user?.nickname,
                 introduction = user?.introduction,
                 profileImageUrl = user?.profileImageUrl
